+ Changed PotentialPayout in Bet.cs from int to double

+++ Bobs-Racing/Models/Bet.cs
@@ -13,7 +13,7 @@ namespace Bobs_Racing.Models
         [Required]
         public int? Amount { get; set; } // Bet amount
 
-        public int PotentialPayout { get; set; } // Potential payout
+        public double PotentialPayout { get; set; } // Potential payout
 
         public bool IsActive { get; set; }
